% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-contactmatrix_list.R
\name{aggregate.contactmatrix_list}
\alias{aggregate.contactmatrix_list}
\title{Aggregate the elements of a \code{contactmatrix_list}}
\usage{
\method{aggregate}{contactmatrix_list}(x, by, FUN = mean, ...)
}
\arguments{
\item{x}{A \code{contactmatrix_list} to aggregate}

\item{by}{Ignored for now}

\item{FUN}{Function to use to aggregate the elements of \code{x} into the output.
Defaults to \code{mean}.}

\item{...}{Addtional arguments passed to \code{FUN}}
}
\value{
A \code{contact_matrix} with the same groupings and same groups (in
practice, same \code{dimnames()}) as all elements of the original
\code{contactmatrix_list} \code{x}.
}
\description{
Aggregate the elements of a \code{contactmatrix_list} into a single
\code{contact_matrix}
}
