% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix.R
\name{assert_contactmatrix}
\alias{assert_contactmatrix}
\title{Assert whether a object is a \strong{valid} \code{contactmatrix} object}
\usage{
assert_contactmatrix(x)
}
\arguments{
\item{x}{object to test}
}
\value{
An error if \code{x} is not a valid \code{contactmatrix} object and \code{x}
invisibly otherwise.
}
\description{
This function asserts that the object \code{x} is a valid \code{contactmatrix}
object.
}
\note{
Checks:
\itemize{
\item \code{x} inherits from the \code{array} class
\item \code{x} inherits from the \code{contactmatrix} class
\item \code{x} contains only numeric values
\item \code{x} is an hypercube, i.e., number of rows and number of columns are equal
\item \code{x} has a \code{symmetric} attribute of class \code{logical}
}
}
\seealso{
\code{\link[=test_contactmatrix_list]{test_contactmatrix_list()}}
}
