% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix_list.R
\name{assert_contactmatrix_list}
\alias{assert_contactmatrix_list}
\title{Assert whether a object is a \strong{valid} \code{contactmatrix_list} object}
\usage{
assert_contactmatrix_list(x)
}
\arguments{
\item{x}{object to test}
}
\value{
An error if \code{x} is not a valid \code{contactmatrix_list} object and \code{x}
invisibly otherwise.
}
\description{
This function asserts that the object \code{x} is a valid \code{contactmatrix_list}
object.
}
\seealso{
\code{\link[=test_contactmatrix_list]{test_contactmatrix_list()}}
}
