% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_get_groupings.R
\name{cm_get_groupings}
\alias{cm_get_groupings}
\title{Get groupings of a \code{contactmatrix} or \code{contactmatrix_list} object}
\usage{
cm_get_groupings(x, ...)
}
\arguments{
\item{x}{A \code{contactmatrix} or \code{contactmatrix_list} object to get groupings
from}

\item{...}{Ignored for now}
}
\description{
Get groupings of a \code{contactmatrix} or \code{contactmatrix_list} object
}
\examples{
cm1 <- new_contactmatrix(
  from  = c("00_05", "05_10", "05_10"),
  to    = c("00_05", "10_15", "15_20"),
  value = c(0.32   , 0.46   , 0.72   )
)

cm2 <- new_contactmatrix(
  from  = c("05_10", "05_10", "05_10"),
  to    = c("00_05", "10_15", "15_20"),
  value = c(0.27   , 0.09   , 0.32   )
)

cml <- new_contactmatrix_list(cm1, cm2)

cm_get_groupings(cm1)
cm_get_groupings(cml)
}
