% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_make_symmetric.R
\name{cm_make_symmetric}
\alias{cm_make_symmetric}
\title{Make a \code{contactmatrix} object symmetric}
\usage{
cm_make_symmetric(x, population)
}
\arguments{
\item{x}{A \code{contactmatrix} object to make symmetric}

\item{population}{A numeric vector of the population size of each group}
}
\value{
A symmetric \code{contactmatrix} object
}
\description{
Symmetricity for social contact matrices is defined as
$c_{ij}N_i == c_{ji}N_j$ where c is the contact value and N the group size.
}
