% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix.R
\docType{package}
\name{contactmatrix}
\alias{contactmatrix}
\alias{contactmatrix-package}
\title{Projected social contact matrices for 152 countries}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
