% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix.R
\name{is_contactmatrix}
\alias{is_contactmatrix}
\title{Test whether a object is a valid \code{contactmatrix} object}
\usage{
is_contactmatrix(x)
}
\arguments{
\item{x}{object to test}
}
\value{
A logical (\code{TRUE} or \code{FALSE}) indicating whether this object is a
contactmatrix object.
}
\description{
This function tests if the object \code{x} inherits from the \code{contactmatrix}
object.
}
\note{
This doesn't say anything about whether \code{x} is a valid \code{contactmatrix} object
as defined in this package. You can use \code{\link[=validate_contactmatrix]{validate_contactmatrix()}} for this.
}
\seealso{
\code{\link[=validate_contactmatrix]{validate_contactmatrix()}}
}
