% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix_list.R
\name{is_contactmatrix_list}
\alias{is_contactmatrix_list}
\title{Test whether a object is a \code{contactmatrix_list} object}
\usage{
is_contactmatrix_list(x)
}
\arguments{
\item{x}{object to test}
}
\value{
A logical (\code{TRUE} or \code{FALSE}) indicating whether this object is a
contactmatrix object.
}
\description{
This function tests if the object \code{x} inherits from the \code{contactmatrix_list}
object.
}
\note{
This doesn't say anything about whether \code{x} is a valid \code{contactmatrix_list}
object as defined in this package. You can use
\code{\link[=assert_contactmatrix_list]{assert_contactmatrix_list()}} for this.
}
\seealso{
\code{\link[=assert_contactmatrix_list]{assert_contactmatrix_list()}}, \code{\link[=test_contactmatrix_list]{test_contactmatrix_list()}}
}
