% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix.R
\name{new_contactmatrix}
\alias{new_contactmatrix}
\title{Constructor for a \code{contactmatrix} object}
\usage{
new_contactmatrix(from, to, value, fill = 0, symmetric = NA)
}
\arguments{
\item{from}{Character vector, or list of named character vectors in the case
of multiple groupings, indicating the characteristics of the group
contacting.}

\item{to}{Character vector, or list of named character vectors in the case
of multiple groupings, indicating the characteristics of the group
contacted.}

\item{value}{Value of the contact between individual in groups \code{from} and
group \code{to}}

\item{fill}{Numeric value to use for contacts non-defined in the \code{from},
\code{to}, \code{value} argument triplet. Defaults to \code{0}; i.e., no contact.}

\item{symmetric}{Logical value (default to \code{NA}) indicating whether the
resulting \code{[}contactmatrix\verb{is symmetric, i.e.,  $c_\{ij\}N_i == c_\{ji\}N_j$ where $i$ is a group from}from\verb{, $j$ a group from }to\verb{, $c$ their corresponding contact }value` and $N$ the group size.}
}
\value{
An array (a matrix if only 1 grouping = 2 dimensions) of subclass
\code{contactmatrix}.
The array has $2n$ dimensions where $n$ is the number of groupings.
}
\description{
Constructor for a \code{contactmatrix} object
}
\details{
\itemize{
\item \code{from}, \code{to} and \code{value} must have the same length
}
}
\examples{
# Usual case; single grouping by age
new_contactmatrix(
  from  = c("00_05", "05_10", "05_10"),
  to    = c("00_05", "10_15", "15_20"),
  value = c(0.32   , 0.46   , 0.72   )
)

# Multiple groupings; gender & case
dat <- data.frame(
  age    = c("young", "young", "old"),
  gender = c("male", "female", "female"),
  age    = c("old", "old", "young"),
  gender = c("female", "female", "female"),
  value  = c(1, 2, 2)
)

new_contactmatrix(
  from  = dat[, c(1, 2)],
  to    = dat[, c(3, 4)],
  value = dat[, "value"]
)

}
