% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix_list.R
\name{new_contactmatrix_list}
\alias{new_contactmatrix_list}
\title{Constructor for a \code{contactmatrix_list} object}
\usage{
new_contactmatrix_list(...)
}
\arguments{
\item{...}{\code{contactmatrix} objects to include in the list. All elements
must have the same groupings.}
}
\value{
A list of \code{contactmatrix} objects, of class \code{contactmatrix_list}.
}
\description{
Constructor for a \code{contactmatrix_list} object
}
\examples{
cm1 <- new_contactmatrix(
  from  = c("00_05", "05_10", "05_10"),
  to    = c("00_05", "10_15", "15_20"),
  value = c(0.32   , 0.46   , 0.72   )
)

cm2 <- new_contactmatrix(
  from  = c("05_10", "05_10", "05_10"),
  to    = c("00_05", "10_15", "15_20"),
  value = c(0.27   , 0.09   , 0.32   )
)

cml <- new_contactmatrix_list(cm1, cm2)
}
