% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-contactmatrix.R
\name{print.contactmatrix}
\alias{print.contactmatrix}
\title{Print a \code{contactmatrix_list}}
\usage{
\method{print}{contactmatrix}(x, ...)
}
\arguments{
\item{x}{A \code{contactmatrix_list} object}

\item{...}{Ignored for now}
}
\value{
\code{x}, invisibly
}
\description{
Print a \code{contactmatrix_list}
}
\examples{
cm2d <- new_contactmatrix(
  from  = c("[0,5)", "[5,10)",  "[5,10)"),
  to    = c("[0,5)", "[10,15)", "[15,20)"),
  value = c(0.32   , 0.46   , 0.72   ),
  symmetric = TRUE
)
print(cm2d)

# Multi-groupings
cm3d <- new_contactmatrix(
  from  = list(
    age = c("young", "young", "old"),
    gender = c("male", "female", "female")
  ),
  to    = list(
    age    = c("old", "old", "young"),
    gender = c("female", "female", "female")
  ),
  value = c(1, 2, 2)
)
print(cm3d)
}
