% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-contactmatrix_list.R
\name{print.contactmatrix_list}
\alias{print.contactmatrix_list}
\title{Print a \code{contactmatrix_list}}
\usage{
\method{print}{contactmatrix_list}(x, ...)
}
\arguments{
\item{x}{A \code{contactmatrix_list} object}

\item{...}{Ignored for now}
}
\value{
\code{x}, invisibly
}
\description{
Print a \code{contactmatrix_list}
}
\examples{
cm1 <- new_contactmatrix(
  from  = c("[0,5)", "[5,10)",  "[5,10)"),
  to    = c("[0,5)", "[10,15)", "[15,20)"),
  value = c(0.32   , 0.46   , 0.72   )
)

cm2 <- new_contactmatrix(
  from  = c("[5,10)", "[5,10)", "[5,10)"),
  to    = c("[0,5)",  "[10,15)", "[10,15)"),
  value = c(0.27   , 0.09   , 0.32   )
)

cml <- new_contactmatrix_list(cm1, cm2)

print(cml)

}
