% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{socialcontacts}
\alias{socialcontacts}
\title{Projected social contact matrices for home, school, work and other locations
for 152 countries}
\format{
A list of with 5 components:
\describe{
  \item{\code{All}}{Social contacts matrix for all locations}
  \item{\code{Home}}{Social contacts matrix for home}
  \item{\code{School}}{Social contacts matrix for school}
  \item{\code{Work}}{Social contacts matrix for work}
  \item{\code{Other}}{Social contacts matrix for other}
}
}
\usage{
socialcontacts
}
\description{
Projected social contact matrices for home, school, work and other locations
for 152 countries
}
\examples{
socialcontacts

}
\keyword{datasets}
