% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix.R
\name{test_contactmatrix}
\alias{test_contactmatrix}
\title{Test whether a object is a \strong{valid} \code{contactmatrix_list} object}
\usage{
test_contactmatrix(x)
}
\arguments{
\item{x}{object to test}
}
\value{
A logical (\code{TRUE} or \code{FALSE}) indicating whether this object is a
valid \code{contactmatrix_list} object.
}
\description{
This function tests if the object \code{x} is a valid \code{contactmatrix_list}
object.
}
\note{
Checks:
\itemize{
\item \code{x} inherits from the \code{array} class
\item \code{x} inherits from the \code{contactmatrix} class
\item \code{x} is an hypercube, i.e., number of rows and number of columns are equal
\item \code{x} has a \code{symmetric} attribute of class \code{logical}
}
}
\seealso{
\code{\link[=test_contactmatrix_list]{test_contactmatrix_list()}}
}
