% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix_list.R
\name{test_contactmatrix_list}
\alias{test_contactmatrix_list}
\title{Test whether a object is a \strong{valid} \code{contactmatrix_list} object}
\usage{
test_contactmatrix_list(x)
}
\arguments{
\item{x}{object to test}
}
\value{
A logical (\code{TRUE} or \code{FALSE}) indicating whether this object is a
valid \code{contactmatrix_list} object.
}
\description{
This function tests if the object \code{x} is a valid \code{contactmatrix_list}
object.
}
\seealso{
\code{\link[=test_contactmatrix_list]{test_contactmatrix_list()}}
}
