% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-contactmatrix_list.R
\name{tidy.contactmatrix_list}
\alias{tidy.contactmatrix_list}
\title{Convert a \code{contactmatrix_list} to a tidy \code{data.frame}}
\usage{
\method{tidy}{contactmatrix_list}(x, ...)
}
\arguments{
\item{x}{A \code{contactmatrix_list} object}

\item{...}{Ignored for now}
}
\description{
Convert a \code{contactmatrix_list} to a tidy \code{data.frame}
}
\examples{
cm1 <- new_contactmatrix(
  from  = c("[0,5)", "[5,10)",  "[5,10)"),
  to    = c("[0,5)", "[10,15)", "[15,20)"),
  value = c(0.32   , 0.46   , 0.72   )
)

cm2 <- new_contactmatrix(
  from  = c("[5,10)",  "[0,5)",   "[5,10)"),
  to    = c("[15,20)", "[10,15)", "[10,15)"),
  value = c(0.27   , 0.09   , 0.32   )
)

cml <- new_contactmatrix_list(cm1, cm2)

tidy(cml)
}
