% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix.R
\name{as_contactmatrix}
\alias{as_contactmatrix}
\alias{as_contactmatrix.default}
\title{Convert an object to a \code{contact_matrix} object.}
\usage{
as_contactmatrix(x, ...)

\method{as_contactmatrix}{default}(x, ...)
}
\arguments{
\item{x}{Object to convert}

\item{...}{Arguments passed to methods}
}
\description{
Convert an object to a \code{contact_matrix} object.
}
