% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_get_groupings.R
\name{cm_get_groupings}
\alias{cm_get_groupings}
\title{Get groupings of a \code{contactmatrix} or \code{contactmatrix_list} object}
\usage{
cm_get_groupings(x, ...)
}
\arguments{
\item{x}{A \code{contactmatrix} or \code{contactmatrix_list} object to get groupings
from}

\item{...}{Ignored for now}
}
\value{
\itemize{
\item In the case of single groupings, a character vector of the groupings
\item In the case of multiple groupings, a list of character vectors of the
groupings
}
}
\description{
Get groupings of a \code{contactmatrix} or \code{contactmatrix_list} object
}
\examples{
cm1 <- new_contactmatrix(
  from  = c("[0,5)", "[5,10)",  "[5,10)"),
  to    = c("[0,5)", "[10,15)", "[15,20)"),
  value = c(0.32   , 0.46   , 0.72   )
)

cm2 <- new_contactmatrix(
  from  = c("[5,10)",  "[0,5)",   "[5,10)"),
  to    = c("[15,20)", "[10,15)", "[10,15)"),
  value = c(0.27   , 0.09   , 0.32   )
)

cml <- new_contactmatrix_list(cm1, cm2)

cm_get_groupings(cm1)
cm_get_groupings(cml)

# Multi-groupings
cm3d <- new_contactmatrix(
  from  = list(
    age = c("young", "young", "old"),
    gender = c("male", "female", "female")
  ),
  to    = list(
    age    = c("old", "old", "young"),
    gender = c("female", "female", "female")
  ),
  value = c(1, 2, 2)
)

cm_get_groupings(cm3d)
}
